Refactored authentication into blueprint
