+++ app/email.py
@@ -34,17 +34,6 @@ def send_mail(subject, sender, recipients, text_body, html_body):
     Thread(target=send_async_email, args=(app, msg)).start()
 
 
-def send_password_reset_email(user):
-    token = user.get_reset_password_token()
-    send_mail('[LA4LD] Reset Your Password',
-              sender=app.config['ADMINS'][0],
-              recipients=[user.email],
-              text_body=render_template('email/reset_password.txt',
-                                        user=user, token=token),
-              html_body=render_template('email/reset_password.html',
-                                        user=user, token=token))
-
-
 def send_new_user_email(user):
     send_mail('[LA4LD] New Account',
               sender=app.config['ADMINS'][0],
