+++ app/routes.py
@@ -7,47 +7,21 @@
 """
 import os
 import json
-from flask import abort, flash, render_template, redirect, url_for, request, \
+from flask import abort, flash, render_template, redirect, url_for,\
     send_from_directory
-from flask_login import current_user, login_user, logout_user, login_required
+from flask_login import current_user, login_required
 from secrets import token_urlsafe
 from werkzeug.utils import secure_filename
-from werkzeug.urls import url_parse
 
 from app import app, db
-from app.email import send_new_user_email, send_password_reset_email
-from app.forms import LoginForm, ResetPasswordRequestForm, ResetPasswordForm,\
-    ImportForm
+from app.email import send_new_user_email
+from app.forms import ImportForm
 from app.models import User, Role
 
 
 @app.route('/')
 def index():
     return render_template('index.html', title='Home Page')
-
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(username=form.username.data).first()
-        if user is None or not user.check_password(form.password.data):
-            flash('Invalid username or password')
-            return redirect(url_for('login'))
-        login_user(user, remember=form.remember_me.data)
-        next_page = request.args.get('next')
-        if not next_page or url_parse(next_page).netloc != '':
-            next_page = url_for('index')
-        return redirect(next_page)
-    return render_template('login.html', title='Sign in', form=form)
-
-
-@app.route('/logout')
-def logout():
-    logout_user()
-    return redirect(url_for('index'))
 
 
 @app.route('/error/test_500')
@@ -60,37 +34,6 @@ def error_test_500():
 def profile():
     user_data = 'abc'
     return render_template('profile.html', title='Profile', data=user_data)
-
-
-@app.route('/reset_password_request', methods=['GET', 'POST'])
-def reset_password_request():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = ResetPasswordRequestForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-        if user:
-            send_password_reset_email(user)
-        flash('Check your email for the instructions to reset your password')
-        return redirect(url_for('login'))
-    return render_template('reset_password_request.html',
-                           title='Reset Password', form=form)
-
-
-@app.route('/reset_password/<token>', methods=['GET', 'POST'])
-def reset_password(token):
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    user = User.verify_reset_password_token(token)
-    if not user:
-        return redirect(url_for('index'))
-    form = ResetPasswordForm()
-    if form.validate_on_submit():
-        user.set_password(form.password.data)
-        db.session.commit()
-        flash('Your password has been reset.')
-        return redirect(url_for('login'))
-    return render_template('reset_password.html', form=form)
 
 
 @app.route('/uploads/<filename>')
