+++ app/templates/auth/login.html
@@ -1,9 +1,15 @@
 {% extends "base.html" %}
 
 {% block app_content %}
-    <h1>Reset Your Password</h1>
     <form action="" method="post">
         {{ form.hidden_tag() }}
+        <p>
+            {{ form.username.label }}<br>
+            {{ form.username(size=32) }}<br>
+            {% for error in form.username.errors %}
+                <span style="color: red;">{{ error }}</span>
+            {% endfor %}
+        </p>
         <p>
             {{ form.password.label }}<br>
             {{ form.password(size=32) }}<br>
@@ -11,13 +17,11 @@
                 <span style="color: red;">{{ error }}</span>
             {% endfor %}
         </p>
-        <p>
-            {{ form.password2.label }}<br>
-            {{ form.password2(size=32) }}<br>
-            {% for error in form.password.errors %}
-                <span style="color: red;">{{ error }}</span>
-            {% endfor %}
-        </p>
+        <p>{{ form.remember_me() }}{{ form.remember_me.label }}</p>
         <p>{{ form.submit() }}</p>
+        <p>
+            Forgot Your Password?
+            <a href="{{ url_for('auth.reset_password_request') }}">Click to Reset It</a>
+        </p>
     </form>
 {% endblock %}
