+++ app/templates/auth/reset_password.html
@@ -4,24 +4,19 @@
     <form action="" method="post">
         {{ form.hidden_tag() }}
         <p>
-            {{ form.username.label }}<br>
-            {{ form.username(size=32) }}<br>
-            {% for error in form.username.errors %}
-                <span style="color: red;">{{ error }}</span>
-            {% endfor %}
-        </p>
-        <p>
             {{ form.password.label }}<br>
             {{ form.password(size=32) }}<br>
             {% for error in form.password.errors %}
                 <span style="color: red;">{{ error }}</span>
             {% endfor %}
         </p>
-        <p>{{ form.remember_me() }}{{ form.remember_me.label }}</p>
-        <p>{{ form.submit() }}</p>
         <p>
-            Forgot Your Password?
-            <a href="{{ url_for('reset_password_request') }}">Click to Reset It</a>
+            {{ form.password2.label }}<br>
+            {{ form.password2(size=32) }}<br>
+            {% for error in form.password.errors %}
+                <span style="color: red;">{{ error }}</span>
+            {% endfor %}
         </p>
+        <p>{{ form.submit() }}</p>
     </form>
 {% endblock %}
