+++ app/templates/auth/reset_password_request.html
@@ -1,7 +1,6 @@
 {% extends "base.html" %}
 
 {% block app_content %}
-    <h1>Reset Password</h1>
     <form action="" method="post">
         {{ form.hidden_tag() }}
         <p>
